python-gen/requirements.txt
reportlab==4.0.4
pdfkit==1.0.0
python-dateutil==2.8.2
Pillow==10.0.0
faker==20.1.0
names==0.3.0
Faker==24.2.0
InvoiceGenerator>=1.1.0
qrcode>=7.3.1 
numpy>=1.24.0 